Add option to schedule passes only within an azimuth window
The user can specify an optional azimuth window in which the antenna can see (useful for fixed stations or obstructions).

If a calculated pass is wholly outside the window then it will be removed. Passes that overlap with the azimuth window will be modified such that only the viewable part of the pass is scheduled. The unviewable/obstructed part of the pass will not be scheduled.

Signed-off-by: Anthony Kremor <satnogs@anthonykremor.com>