Introduce new method for satellite search
The new method uses only inputs which are the direct results of API
queries on SatNOGS DB and Network, without any logic applied. This
simplifies the caching logic.

The previous method of searching for satellites was filtering the list
of transmitter twice: Once when creating the transmitters_{id}.txt file
and after reading this file.

Signed-off-by: Fabian P. Schmidt <kerel@mailbox.org>